pyproject.toml
[project]
name = "PyTango-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "gevent>=25.9.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pytango==10.1.2",
]
